Code refactoring

+++ routes/web.php
@@ -1,20 +1,19 @@
 <?php
 
-Route::get('/', 'HomeController@home');
+// Admin Routes
+Route::get('admin/products', 'Admin\ProductController@index')->middleware('auth');
+Route::get('admin/products/create',   'Admin\ProductController@create')->middleware('auth'); // Show form to add new product
+Route::post('admin/products',         'Admin\ProductController@store')->middleware('auth'); // Store the product in database
+Route::get('admin/products/{id}/edit','Admin\ProductController@edit')->middleware('auth'); // Show form to edit a specific existing product
+Route::patch('admin/products/{id}',   'Admin\ProductController@update')->middleware('auth'); // update product details in database
+Route::delete('admin/products/{id}',  'Admin\ProductController@destroy')->middleware('auth'); // delete a record from database
 
-Route::get('/products',          'ProductController@index'); // List all products
-Route::get('/products/create',   'ProductController@create')->middleware('auth'); // Show form to add new product
-Route::post('/products',         'ProductController@store')->middleware('auth'); // Store the product in database
-Route::get('/products/{id}',     'ProductController@show'); // Show one product
-Route::get('/products/{id}/edit','ProductController@edit')->middleware('auth'); // Show form to edit a specific existing product
-Route::patch('/products/{id}',   'ProductController@update')->middleware('auth'); // update product details in database
-Route::delete('/products/{id}',  'ProductController@destroy')->middleware('auth'); // delete a record from database
+Route::get('admin/categories', 'Admin\CategoryController@index')->middleware('auth'); // Show form to add new category
+Route::get('admin/categories/create', 'Admin\CategoryController@create')->middleware('auth'); // Show form to add new category
+Route::post('admin/categories', 'Admin\CategoryController@store')->middleware('auth'); // Store the submitted category in database
 
-Route::get('/categories/create', 'CategoryController@create')->middleware('auth'); // Show form to add new category
-Route::post('/categories', 'CategoryController@store')->middleware('auth'); // Store the submitted category in database
-
-Route::get('/orders', 'OrderController@index');//->middleware('auth');
-Route::get('/orders/{id}', 'OrderController@show');//->middleware('auth');
+Route::get('admin/orders', 'Admin\OrderController@index')->middleware('auth');
+Route::get('admin/orders/{id}', 'Admin\OrderController@show')->middleware('auth');
 
 Auth::routes(['register' => false]);
 
@@ -25,4 +24,9 @@ Route::get('/customer/login', 'Customer\AuthController@showLoginForm');
 Route::post('/customer/login', 'Customer\AuthController@login');
 Route::post('/customer/logout', 'Customer\AuthController@logoutt');
 
-Route::post('products/{product_id}/ratings', 'ProductRatingController@store');
+Route::post('customer/products/{product_id}/ratings', 'Customer\ProductRatingController@store');
+
+Route::get('/', 'HomeController@home');
+
+Route::get('customer/products',          'Customer\ProductController@index'); // List all products
+Route::get('customer/products/{id}',     'Customer\ProductController@show'); // Show one product
